Significantly optimize SpatialHashingSystem, BoidSystem and TurretSystem
SpatialHashingSystem computes the grid in a new way which is faster, and stores velocity and positions directly in the grid alongside references to entities. This lets the boid system directly read the information it needs in contiguous memory instead of having to do random access lookup from the entity, significantly reducing cache misses. We also compute a second datastructure which stores number of birds in each cell. This lets the turret system determine a good target much faster.

Boid system has reduced number of square roots by merging multiple vectors before steertowards. There is simplified logic when counting neighbours that only requires quering the spatial datastructure once instead of twice, and which does not use randomness and nested loops which are bad for cpu branch prediction

The turret system has been moved to jobs, the targetting is simplified as it can simply sample the datastructure produced by spatialHashingSystem and the quaternion math has been changed to use dots types so it can be burst compiled

Explosion system logic has been inverted so the boid system listens for explosions and reacts instead of explosions iterating over all boids